Updated the way complex connections are specified
Connections which rely on properties in other reservoirs (e.g., scale
a flux in response to the mass in the upstream reservoir) now require
that the connection-type is specified via the ctype keyword. This
increases the readability of the model setup, and simplifies code
maintenance.

I also added new functions to list all connection objects associated
with a reservoir, and to list all processes associated with a
connection. This simplified connection introspection

Updates to the quickstart guide which reflect these changes