updated cs2 to compute
not only the burial flux, but the actual flux of DIC and TA going into, or leaving
the deep box. This requires that the flux export from the surface box, does bypass
the sink, i.e., it does not actively add DIC and TA to the sink. Use the bypass="sink"
parameter when specifying particulate carbonate export fluxes!