save_flux now saves the data back onto the flux object
However: This will only save values for the first n-steps of the solver. The remainder of the flux data remains at zero. I.e., there is not yet a mapping of flux data to time

This will only work if the solver progresses unidirectional through time. Solver that go forward and backward may  show weird data

This currently only works for the flux mass. Isotopes are not implemented yet.