feat: Add register_return_values function
This commit adds a new function `register_return_values` to the `utility_functions.py` file. This function is used to check the return values of external function instances and create the necessary reservoirs. The function takes an `ec` parameter and a `parent` parameter. The `ec` parameter represents the external function instances and the `parent` parameter is used to register the reservoirs. The function iterates over the return values of `ec` and creates reservoirs based on the type of the value. If the value is a dictionary, a new reservoir is created with the name, species, concentration, register, volume, and rtype values extracted from the dictionary. If the value is a `Reservoir` object, the `ef_results` attribute of the `Reservoir` object is set to `True`.