feat(esbmtk): add support for creating reservoirs or fluxes
from return values defined in external_function definitions

This commit adds support for creating reservoirs or fluxes from the return values of external function instances in the `register_return_values` function of the `esbmtk.utility_functions` module. Reservoirs are created for return values that are not dictionaries, while reservoirs and fluxes are created for return values that are dictionaries. This allows for more flexibility in handling return values and improves the functionality of the module.